Bring back action_probs and multiclassprobs

+++ vowpalwabbit/classweight.cc
@@ -52,10 +52,10 @@ static void predict_or_learn(classweights& cweights, LEARNER::single_learner& ba
 {
   switch (pred_type)
   {
-    case static_cast<int>(prediction_type_t::scalar):
+    case prediction_type_t::scalar:
       ec.weight *= cweights.get_class_weight((uint32_t)ec.l.simple().label);
       break;
-    case static_cast<int>(prediction_type_t::multiclass):
+    case prediction_type_t::multiclass:
       ec.weight *= cweights.get_class_weight(ec.l.multi().label);
       break;
     default:
@@ -93,14 +93,14 @@ LEARNER::base_learner* classweight_setup(options_i& options, vw& all)
   LEARNER::learner<classweights, example>* ret;
   if (base->pred_type == prediction_type_t::scalar)
   {
-    ret = &LEARNER::init_learner<classweights>(cweights, base, predict_or_learn<true, static_cast<int>(prediction_type_t::scalar)>,
-        predict_or_learn<false, static_cast<int>(prediction_type_t::scalar)>);
+    ret = &LEARNER::init_learner<classweights>(cweights, base, predict_or_learn<true, prediction_type_t::scalar>,
+        predict_or_learn<false, prediction_type_t::scalar>);
     ret->label_type = label_type_t::simple;
   }
   else if (base->pred_type == prediction_type_t::multiclass)
   {
-    ret = &LEARNER::init_learner<classweights>(cweights, base, predict_or_learn<true, static_cast<int>(prediction_type_t::multiclass)>,
-        predict_or_learn<false, static_cast<int>(prediction_type_t::multiclass)>);
+    ret = &LEARNER::init_learner<classweights>(cweights, base, predict_or_learn<true, prediction_type_t::multiclass>,
+        predict_or_learn<false, prediction_type_t::multiclass>);
     ret->label_type = label_type_t::multi;
   }
   else
